Create Task View
Working on the functionality of the task view where user can edit task title and description and can add subtask and description. Also add the delete task function and a function that move the task to the chosen column and delete its data from previous. Still working on the light and theme design.